Start tracking Branded.xccheckout file
Not sure why this file wasn't created / tracked when
the project was created originally. I tried to remove this
file, but Xcode keeps re-generating it. Since this file
is in xcsharedata directory and none of my personal Xcode
related user data is included in this file there is no harm
in tracking it.