Create fake menu reader as a stand-in for a real reader
CCHFakeMenuItemsReader is used only in tests. Using a fake
object (or test double) to play the CCHMenuItemsReader role
instead of a real menu reader object has several benefits:

1. Test doubles are very easy to make. Nothing hinders
you from creating one for every possible situation. Each
variation is like an artist’s sketch. It emphasizes a single
interesting feature and allows the underlying object’s
other details to recede to the background.

2. We can return a canned answer (or value) from test
doubles. It might look like these test doubles are quite
limiting in terms of functionality, but that's the whole
point. The fact that they always return a pre-determined
value is perfect. This stubbed return value provides a
dependable foundation on which to construct the test.

3. Using test doubles to inject dependencies helps us
decouple the test from actual class. It no longer matters
if CCHMenuItemsPlistReader (the real menu reader object)
is slow because CCHFakeMenuItemsReader is always fast.

As a result, the test doubles are simple, fast, isolated
and intention revealing during testing.

For more information on this topic, please refer to the
“Testing Incoming Messages” section from Chapter 9
of Practical Object-Oriented Design in Ruby book:
http://goo.gl/NZ9Isu